Removed depracated PE class, added option to module overloading on whether to require signed binaries, updated Mimikatz loader to use DInvoke and module overloading
