+++ SharpSploit/Execution/ManualMap/Map.cs
@@ -7,6 +7,381 @@ using Execute = SharpSploit.Execution;
 
 namespace SharpSploit.Execution.ManualMap
 {
+    /// <summary>
+    /// Holds data structures for using PEs.
+    /// </summary>
+    public class PE
+    {
+        // DllMain constants
+        public const UInt32 DLL_PROCESS_DETACH = 0;
+        public const UInt32 DLL_PROCESS_ATTACH = 1;
+        public const UInt32 DLL_THREAD_ATTACH = 2;
+        public const UInt32 DLL_THREAD_DETACH = 3;
+
+        // Primary class for loading PE
+        [UnmanagedFunctionPointer(CallingConvention.StdCall)]
+        public delegate bool DllMain(IntPtr hinstDLL, uint fdwReason, IntPtr lpvReserved);
+
+        [Flags]
+        public enum DataSectionFlags : uint
+        {
+            TYPE_NO_PAD = 0x00000008,
+            CNT_CODE = 0x00000020,
+            CNT_INITIALIZED_DATA = 0x00000040,
+            CNT_UNINITIALIZED_DATA = 0x00000080,
+            LNK_INFO = 0x00000200,
+            LNK_REMOVE = 0x00000800,
+            LNK_COMDAT = 0x00001000,
+            NO_DEFER_SPEC_EXC = 0x00004000,
+            GPREL = 0x00008000,
+            MEM_FARDATA = 0x00008000,
+            MEM_PURGEABLE = 0x00020000,
+            MEM_16BIT = 0x00020000,
+            MEM_LOCKED = 0x00040000,
+            MEM_PRELOAD = 0x00080000,
+            ALIGN_1BYTES = 0x00100000,
+            ALIGN_2BYTES = 0x00200000,
+            ALIGN_4BYTES = 0x00300000,
+            ALIGN_8BYTES = 0x00400000,
+            ALIGN_16BYTES = 0x00500000,
+            ALIGN_32BYTES = 0x00600000,
+            ALIGN_64BYTES = 0x00700000,
+            ALIGN_128BYTES = 0x00800000,
+            ALIGN_256BYTES = 0x00900000,
+            ALIGN_512BYTES = 0x00A00000,
+            ALIGN_1024BYTES = 0x00B00000,
+            ALIGN_2048BYTES = 0x00C00000,
+            ALIGN_4096BYTES = 0x00D00000,
+            ALIGN_8192BYTES = 0x00E00000,
+            ALIGN_MASK = 0x00F00000,
+            LNK_NRELOC_OVFL = 0x01000000,
+            MEM_DISCARDABLE = 0x02000000,
+            MEM_NOT_CACHED = 0x04000000,
+            MEM_NOT_PAGED = 0x08000000,
+            MEM_SHARED = 0x10000000,
+            MEM_EXECUTE = 0x20000000,
+            MEM_READ = 0x40000000,
+            MEM_WRITE = 0x80000000
+        }
+
+
+        public struct IMAGE_DOS_HEADER
+        {      // DOS .EXE header
+            public UInt16 e_magic;              // Magic number
+            public UInt16 e_cblp;               // Bytes on last page of file
+            public UInt16 e_cp;                 // Pages in file
+            public UInt16 e_crlc;               // Relocations
+            public UInt16 e_cparhdr;            // Size of header in paragraphs
+            public UInt16 e_minalloc;           // Minimum extra paragraphs needed
+            public UInt16 e_maxalloc;           // Maximum extra paragraphs needed
+            public UInt16 e_ss;                 // Initial (relative) SS value
+            public UInt16 e_sp;                 // Initial SP value
+            public UInt16 e_csum;               // Checksum
+            public UInt16 e_ip;                 // Initial IP value
+            public UInt16 e_cs;                 // Initial (relative) CS value
+            public UInt16 e_lfarlc;             // File address of relocation table
+            public UInt16 e_ovno;               // Overlay number
+            public UInt16 e_res_0;              // Reserved words
+            public UInt16 e_res_1;              // Reserved words
+            public UInt16 e_res_2;              // Reserved words
+            public UInt16 e_res_3;              // Reserved words
+            public UInt16 e_oemid;              // OEM identifier (for e_oeminfo)
+            public UInt16 e_oeminfo;            // OEM information; e_oemid specific
+            public UInt16 e_res2_0;             // Reserved words
+            public UInt16 e_res2_1;             // Reserved words
+            public UInt16 e_res2_2;             // Reserved words
+            public UInt16 e_res2_3;             // Reserved words
+            public UInt16 e_res2_4;             // Reserved words
+            public UInt16 e_res2_5;             // Reserved words
+            public UInt16 e_res2_6;             // Reserved words
+            public UInt16 e_res2_7;             // Reserved words
+            public UInt16 e_res2_8;             // Reserved words
+            public UInt16 e_res2_9;             // Reserved words
+            public UInt32 e_lfanew;             // File address of new exe header
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct IMAGE_DATA_DIRECTORY
+        {
+            public UInt32 VirtualAddress;
+            public UInt32 Size;
+        }
+
+        [StructLayout(LayoutKind.Sequential, Pack = 1)]
+        public struct IMAGE_OPTIONAL_HEADER32
+        {
+            public UInt16 Magic;
+            public Byte MajorLinkerVersion;
+            public Byte MinorLinkerVersion;
+            public UInt32 SizeOfCode;
+            public UInt32 SizeOfInitializedData;
+            public UInt32 SizeOfUninitializedData;
+            public UInt32 AddressOfEntryPoint;
+            public UInt32 BaseOfCode;
+            public UInt32 BaseOfData;
+            public UInt32 ImageBase;
+            public UInt32 SectionAlignment;
+            public UInt32 FileAlignment;
+            public UInt16 MajorOperatingSystemVersion;
+            public UInt16 MinorOperatingSystemVersion;
+            public UInt16 MajorImageVersion;
+            public UInt16 MinorImageVersion;
+            public UInt16 MajorSubsystemVersion;
+            public UInt16 MinorSubsystemVersion;
+            public UInt32 Win32VersionValue;
+            public UInt32 SizeOfImage;
+            public UInt32 SizeOfHeaders;
+            public UInt32 CheckSum;
+            public UInt16 Subsystem;
+            public UInt16 DllCharacteristics;
+            public UInt32 SizeOfStackReserve;
+            public UInt32 SizeOfStackCommit;
+            public UInt32 SizeOfHeapReserve;
+            public UInt32 SizeOfHeapCommit;
+            public UInt32 LoaderFlags;
+            public UInt32 NumberOfRvaAndSizes;
+
+            public IMAGE_DATA_DIRECTORY ExportTable;
+            public IMAGE_DATA_DIRECTORY ImportTable;
+            public IMAGE_DATA_DIRECTORY ResourceTable;
+            public IMAGE_DATA_DIRECTORY ExceptionTable;
+            public IMAGE_DATA_DIRECTORY CertificateTable;
+            public IMAGE_DATA_DIRECTORY BaseRelocationTable;
+            public IMAGE_DATA_DIRECTORY Debug;
+            public IMAGE_DATA_DIRECTORY Architecture;
+            public IMAGE_DATA_DIRECTORY GlobalPtr;
+            public IMAGE_DATA_DIRECTORY TLSTable;
+            public IMAGE_DATA_DIRECTORY LoadConfigTable;
+            public IMAGE_DATA_DIRECTORY BoundImport;
+            public IMAGE_DATA_DIRECTORY IAT;
+            public IMAGE_DATA_DIRECTORY DelayImportDescriptor;
+            public IMAGE_DATA_DIRECTORY CLRRuntimeHeader;
+            public IMAGE_DATA_DIRECTORY Reserved;
+        }
+
+        [StructLayout(LayoutKind.Sequential, Pack = 1)]
+        public struct IMAGE_OPTIONAL_HEADER64
+        {
+            public UInt16 Magic;
+            public Byte MajorLinkerVersion;
+            public Byte MinorLinkerVersion;
+            public UInt32 SizeOfCode;
+            public UInt32 SizeOfInitializedData;
+            public UInt32 SizeOfUninitializedData;
+            public UInt32 AddressOfEntryPoint;
+            public UInt32 BaseOfCode;
+            public UInt64 ImageBase;
+            public UInt32 SectionAlignment;
+            public UInt32 FileAlignment;
+            public UInt16 MajorOperatingSystemVersion;
+            public UInt16 MinorOperatingSystemVersion;
+            public UInt16 MajorImageVersion;
+            public UInt16 MinorImageVersion;
+            public UInt16 MajorSubsystemVersion;
+            public UInt16 MinorSubsystemVersion;
+            public UInt32 Win32VersionValue;
+            public UInt32 SizeOfImage;
+            public UInt32 SizeOfHeaders;
+            public UInt32 CheckSum;
+            public UInt16 Subsystem;
+            public UInt16 DllCharacteristics;
+            public UInt64 SizeOfStackReserve;
+            public UInt64 SizeOfStackCommit;
+            public UInt64 SizeOfHeapReserve;
+            public UInt64 SizeOfHeapCommit;
+            public UInt32 LoaderFlags;
+            public UInt32 NumberOfRvaAndSizes;
+
+            public IMAGE_DATA_DIRECTORY ExportTable;
+            public IMAGE_DATA_DIRECTORY ImportTable;
+            public IMAGE_DATA_DIRECTORY ResourceTable;
+            public IMAGE_DATA_DIRECTORY ExceptionTable;
+            public IMAGE_DATA_DIRECTORY CertificateTable;
+            public IMAGE_DATA_DIRECTORY BaseRelocationTable;
+            public IMAGE_DATA_DIRECTORY Debug;
+            public IMAGE_DATA_DIRECTORY Architecture;
+            public IMAGE_DATA_DIRECTORY GlobalPtr;
+            public IMAGE_DATA_DIRECTORY TLSTable;
+            public IMAGE_DATA_DIRECTORY LoadConfigTable;
+            public IMAGE_DATA_DIRECTORY BoundImport;
+            public IMAGE_DATA_DIRECTORY IAT;
+            public IMAGE_DATA_DIRECTORY DelayImportDescriptor;
+            public IMAGE_DATA_DIRECTORY CLRRuntimeHeader;
+            public IMAGE_DATA_DIRECTORY Reserved;
+        }
+
+        [StructLayout(LayoutKind.Sequential, Pack = 1)]
+        public struct IMAGE_FILE_HEADER
+        {
+            public UInt16 Machine;
+            public UInt16 NumberOfSections;
+            public UInt32 TimeDateStamp;
+            public UInt32 PointerToSymbolTable;
+            public UInt32 NumberOfSymbols;
+            public UInt16 SizeOfOptionalHeader;
+            public UInt16 Characteristics;
+        }
+
+        [StructLayout(LayoutKind.Explicit)]
+        public struct IMAGE_SECTION_HEADER
+        {
+            [FieldOffset(0)]
+            [MarshalAs(UnmanagedType.ByValArray, SizeConst = 8)]
+            public char[] Name;
+            [FieldOffset(8)]
+            public UInt32 VirtualSize;
+            [FieldOffset(12)]
+            public UInt32 VirtualAddress;
+            [FieldOffset(16)]
+            public UInt32 SizeOfRawData;
+            [FieldOffset(20)]
+            public UInt32 PointerToRawData;
+            [FieldOffset(24)]
+            public UInt32 PointerToRelocations;
+            [FieldOffset(28)]
+            public UInt32 PointerToLinenumbers;
+            [FieldOffset(32)]
+            public UInt16 NumberOfRelocations;
+            [FieldOffset(34)]
+            public UInt16 NumberOfLinenumbers;
+            [FieldOffset(36)]
+            public DataSectionFlags Characteristics;
+
+            public string Section
+            {
+                get { return new string(Name); }
+            }
+        }
+
+        [StructLayout(LayoutKind.Explicit)]
+        public struct IMAGE_EXPORT_DIRECTORY
+        {
+            [FieldOffset(0)]
+            public UInt32 Characteristics;
+            [FieldOffset(4)]
+            public UInt32 TimeDateStamp;
+            [FieldOffset(8)]
+            public UInt16 MajorVersion;
+            [FieldOffset(10)]
+            public UInt16 MinorVersion;
+            [FieldOffset(12)]
+            public UInt32 Name;
+            [FieldOffset(16)]
+            public UInt32 Base;
+            [FieldOffset(20)]
+            public UInt32 NumberOfFunctions;
+            [FieldOffset(24)]
+            public UInt32 NumberOfNames;
+            [FieldOffset(28)]
+            public UInt32 AddressOfFunctions;
+            [FieldOffset(32)]
+            public UInt32 AddressOfNames;
+            [FieldOffset(36)]
+            public UInt32 AddressOfOrdinals;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct IMAGE_BASE_RELOCATION
+        {
+            public uint VirtualAdress;
+            public uint SizeOfBlock;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct PE_META_DATA
+        {
+            public UInt32 Pe;
+            public Boolean Is32Bit;
+            public IMAGE_FILE_HEADER ImageFileHeader;
+            public IMAGE_OPTIONAL_HEADER32 OptHeader32;
+            public IMAGE_OPTIONAL_HEADER64 OptHeader64;
+            public IMAGE_SECTION_HEADER[] Sections;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct PE_MANUAL_MAP
+        {
+            public String DecoyModule;
+            public IntPtr ModuleBase;
+            public PE_META_DATA PEINFO;
+        }
+
+        [StructLayout(LayoutKind.Explicit)]
+        public struct IMAGE_THUNK_DATA32
+        {
+            [FieldOffset(0)]
+            public UInt32 ForwarderString;
+            [FieldOffset(0)]
+            public UInt32 Function;
+            [FieldOffset(0)]
+            public UInt32 Ordinal;
+            [FieldOffset(0)]
+            public UInt32 AddressOfData;
+        }
+
+        [StructLayout(LayoutKind.Explicit)]
+        public struct IMAGE_THUNK_DATA64
+        {
+            [FieldOffset(0)]
+            public UInt64 ForwarderString;
+            [FieldOffset(0)]
+            public UInt64 Function;
+            [FieldOffset(0)]
+            public UInt64 Ordinal;
+            [FieldOffset(0)]
+            public UInt64 AddressOfData;
+        }
+
+        [StructLayout(LayoutKind.Explicit)]
+        public struct ApiSetNamespace
+        {
+            [FieldOffset(0x0C)]
+            public int Count;
+
+            [FieldOffset(0x10)]
+            public int EntryOffset;
+        }
+
+        [StructLayout(LayoutKind.Explicit, Size = 24)]
+        public struct ApiSetNamespaceEntry
+        {
+            [FieldOffset(0x04)]
+            public int NameOffset;
+
+            [FieldOffset(0x08)]
+            public int NameLength;
+
+            [FieldOffset(0x10)]
+            public int ValueOffset;
+        }
+
+        [StructLayout(LayoutKind.Explicit)]
+        public struct ApiSetValueEntry
+        {
+            [FieldOffset(0x0C)]
+            public int ValueOffset;
+
+            [FieldOffset(0x10)]
+            public int ValueCount;
+        }
+
+        [StructLayout(LayoutKind.Sequential)]
+        public struct LDR_DATA_TABLE_ENTRY
+        {
+            public Native.LIST_ENTRY InLoadOrderLinks;
+            public Native.LIST_ENTRY InMemoryOrderLinks;
+            public Native.LIST_ENTRY InInitializationOrderLinks;
+            public IntPtr DllBase;
+            public IntPtr EntryPoint;
+            public UInt32 SizeOfImage;
+            public Native.UNICODE_STRING FullDllName;
+            public Native.UNICODE_STRING BaseDllName;
+        }
+    }//end class
+
+    /// <summary>
+    /// Class for manually mapping PEs.
+    /// </summary>
     public class Map
     {
